+++ lib/ws_servers/api/handlers/send_settings.js
@@ -1,9 +1,9 @@
 'use strict'
 
-const send = require('../../util/ws/send')
-const sendError = require('../../util/ws/send_error')
-const isAuthorized = require('../../util/ws/is_authorized')
-const getUserSettings = require('../../util/user_settings')
+const send = require('../../../util/ws/send')
+const sendError = require('../../../util/ws/send_error')
+const isAuthorized = require('../../../util/ws/is_authorized')
+const getUserSettings = require('../../../util/user_settings')
 
 module.exports = async (server, ws, msg) => {
   const { db } = server
