+++ lib/ws_servers/api/index.js
@@ -22,7 +22,7 @@ const onAlgoOrderSubmit = require('./handlers/on_algo_order_submit')
 const onAlgoOrderCancel = require('./handlers/on_algo_order_cancel')
 const onAlgoPause = require('./handlers/on_algo_pause')
 const onSettingsUpdate = require('./handlers/on_settings_update')
-const onSettingsRequest = require('./send_settings')
+const onSettingsRequest = require('./handlers/send_settings')
 const onCoreSettingsRequest = require('./handlers/on_core_settings_request')
 const onOrderHistoryRequest = require('./handlers/on_order_history_request')
 const onPastStrategiesRequest = require('./handlers/strategy/past_strategies_request')
